+++ PKHeX.Core/PKM/PKM.cs
@@ -756,7 +756,7 @@ namespace PKHeX.Core
         {
             AbilityNumber = 1 << n;
             int[] abilities = PersonalInfo.Abilities;
-            if (n < abilities.Length)
+            if ((uint)n < abilities.Length)
             {
                 if (abilities[n] == abilities[0])
                     n = 0;
